reverted previous changes, from now: - on startup if price changes server will mark a house for update (then remove the update check if a house has owner what is checked in later phase of loading), then after all items load, checked house doors will update with new price - when reloading, all houses will be updated with new prices/ rents together with doors (if there is no owner)
This commit, as previous (a wannabe) tries to fix the bug with incorrect house prices on doors. It requires so much stuff due to reloading as well as TFS advanced options in house renting/ buying stuff.
Please correct me if I've done any bugs or you don't get the logic and you got the working one.

git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@3723 c161f52b-5830-4e7a-bee0-5dea2c17262e